moved calendar related dialogs to new architecture package
